+++ back/app/services/projects_filtering_service.rb
@@ -1,20 +1,16 @@
 class ProjectsFilteringService
   include Filterer
 
-  add_filter("by_topics") do |scope, options|
+  add_filter('by_topics') do |scope, options|
     topics = options[:topics]
-    topics ? scope.with_all_topics(topics) : scope
+    topics ? scope.with_some_topics(topics) : scope
   end
 
-  add_filter("by_areas") do |scope, options|
-    if (areas = options[:areas])
-      scope.with_some_areas(areas)
-    else
-      scope
-    end
+  add_filter('by_areas') do |scope, options|
+    options[:areas] ? scope.with_some_areas(options[:areas]) : scope
   end
 
-  add_filter("filter_ids") do |scope, options|
+  add_filter('filter_ids') do |scope, options|
     keep_ids = options[:filter_ids]
     keep_ids ? scope.where(id: keep_ids) : scope
   end